web_app/requirements.txt
pandas==1.0.2
tensorflow==2.3.1
streamlit==0.56.0
numpy==1.18.1
Pillow==8.0.1
